Fixed: Aterado o tipo Integer para Long no atributo rg da classe Pessoa, pois estava causando um erro ao inserir algumas possibilidades de numeros de documentos

+++ src/main/java/br/unisinos/estruturas/trabalho/gb/entity/Pessoa.java
@@ -6,12 +6,12 @@ public class Pessoa {
 
     private String nome;
     private String cpf;
-    private Integer rg;
+    private Long rg;
 
     private LocalDate data;
     private String cidade;
 
-    public Pessoa(String nome, String cpf, Integer rg, LocalDate data, String cidade) {
+    public Pessoa(String nome, String cpf, Long rg, LocalDate data, String cidade) {
         this.nome = nome;
         this.cpf = cpf;
         this.rg = rg;
@@ -27,7 +27,7 @@ public class Pessoa {
         return cpf;
     }
 
-    public Integer getRg() {
+    public Long getRg() {
         return rg;
     }
